backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
python-multipart
langchain
langchain-community
langchain-google-genai
google-generativeai
pypdf
faiss-cpu
sentence-transformers